+++ Classes/Collection/XmlNamespace.php
@@ -18,4 +18,5 @@ namespace Brotkrueml\FeedGenerator\Collection;
 enum XmlNamespace: string
 {
     case atom = 'http://www.w3.org/2005/Atom';
+    case content = 'http://purl.org/rss/1.0/modules/content/';
 }